added a create page and an edit page
