+++ manage-project-frontend/src/App.js
@@ -1,5 +1,5 @@
 import Header from './components/Header'
-import Main from './components/Main';
+import Main from './components/Main'
 import './styles.scss'
 
 function App() {
