+++ manage-project-frontend/src/components/Header.js
@@ -1,4 +1,4 @@
-import { Link } from 'react-router-dom';
+import { Link } from 'react-router-dom'
 
 const Header = (props) => {
   return (
